:recycle: Move docker view to drf-standardized-errors
