+++ backend/zane_api/views/docker.py
@@ -19,52 +19,32 @@ class DockerImageSerializer(serializers.Serializer):
     description = serializers.CharField()
 
 
-class DockerSuccessResponseSerializer(serializers.Serializer):
+class DockerImageSearchResponseSerializer(serializers.Serializer):
     images = DockerImageSerializer(many=True)
 
 
-class DockerImageListSearchSerializer(serializers.Serializer):
+class DockerImageSearchParamsSerializer(serializers.Serializer):
     q = serializers.CharField(required=True)
 
 
-class DockerImageSearchErrorSerializer(serializers.BaseErrorSerializer):
-    q = serializers.StringListField(required=False)
-
-
-class DockerImageSearchErrorResponseSerializer(serializers.Serializer):
-    errors = DockerImageSearchErrorSerializer()
-
-
 class DockerImageSearchView(APIView):
-    serializer_class = DockerSuccessResponseSerializer
-    forbidden_serializer_class = serializers.ForbiddenResponseSerializer
-    error_serializer_class = DockerImageSearchErrorResponseSerializer
+    serializer_class = DockerImageSearchResponseSerializer
 
     @extend_schema(
         parameters=[
-            DockerImageListSearchSerializer,
+            DockerImageSearchParamsSerializer,
         ],
-        responses={
-            200: serializer_class,
-            403: forbidden_serializer_class,
-            422: error_serializer_class,
-        },
         operation_id="searchDockerRegistry",
     )
     def get(self, request: Request):
         query_params = request.query_params.dict()
-        form = DockerImageListSearchSerializer(data=query_params)
+        form = DockerImageSearchParamsSerializer(data=query_params)
 
-        if form.is_valid():
+        if form.is_valid(raise_exception=True):
             params = form.data
             result = search_images_docker_hub(term=params["q"])
-            response = self.serializer_class({"images": result})
+            response = DockerImageSearchResponseSerializer({"images": result})
             return Response(response.data, status=status.HTTP_200_OK)
-
-        return Response(
-            status=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            data={"errors": form.errors},
-        )
 
 
 class DockerLoginSuccessResponseSerializer(serializers.Serializer):
@@ -77,16 +57,6 @@ class DockerLoginRequestSerializer(serializers.Serializer):
     registry_url = serializers.URLField(required=False)
 
 
-class DockerLoginErrorSerializer(serializers.BaseErrorSerializer):
-    username = serializers.StringListField(required=False)
-    password = serializers.StringListField(required=False)
-    registry_url = serializers.StringListField(required=False)
-
-
-class DockerLoginErrorResponseSerializer(serializers.Serializer):
-    errors = DockerLoginErrorSerializer()
-
-
 class DockerLoginView(APIView):
     serializer_class = DockerLoginSuccessResponseSerializer
 
@@ -97,7 +67,7 @@ class DockerLoginView(APIView):
     def post(self, request: Request):
         form = DockerLoginRequestSerializer(data=request.data)
 
-        if form.is_valid():
+        if form.is_valid(raise_exception=True):
             data = form.data
             try:
                 login_to_docker_registry(**data)
@@ -107,13 +77,8 @@ class DockerLoginView(APIView):
                 response = self.serializer_class({"success": True})
                 return Response(response.data, status=status.HTTP_200_OK)
 
-        return Response(
-            status=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            data={"errors": form.errors},
-        )
 
-
-class DockerPortCheckSuccessResponseSerializer(serializers.Serializer):
+class DockerPortCheckResponseSerializer(serializers.Serializer):
     available = serializers.BooleanField()
 
 
@@ -121,43 +86,22 @@ class DockerPortCheckRequestSerializer(serializers.Serializer):
     port = serializers.IntegerField(required=True, min_value=0)
 
 
-class DockerPortCheckErrorSerializer(serializers.BaseErrorSerializer):
-    port = serializers.StringListField(required=False)
-
-
-class DockerPortCheckErrorResponseSerializer(serializers.Serializer):
-    errors = DockerPortCheckErrorSerializer()
-
-
 class DockerPortCheckView(APIView):
-    serializer_class = DockerPortCheckSuccessResponseSerializer
-    forbidden_serializer_class = serializers.ForbiddenResponseSerializer
-    error_serializer_class = DockerPortCheckErrorResponseSerializer
+    serializer_class = DockerPortCheckResponseSerializer
 
     @extend_schema(
         request=DockerPortCheckRequestSerializer,
-        responses={
-            200: serializer_class,
-            400: serializer_class,
-            403: forbidden_serializer_class,
-            422: error_serializer_class,
-        },
         operation_id="checkIfPortIsAvailable",
     )
     def post(self, request: Request):
         form = DockerPortCheckRequestSerializer(data=request.data)
 
-        if form.is_valid():
+        if form.is_valid(raise_exception=True):
             data = form.data
             result = check_if_port_is_available_on_host(port=data["port"])
 
-            response = self.serializer_class({"available": result})
+            response = DockerPortCheckResponseSerializer({"available": result})
             return Response(
                 response.data,
-                status=status.HTTP_200_OK if result else status.HTTP_400_BAD_REQUEST,
+                status=status.HTTP_200_OK,
             )
-
-        return Response(
-            status=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            data={"errors": form.errors},
-        )
